docs: updates NLP Manager documentation
I noticed that the Transformers documentation was incorrect, using
`transformer` instead of `processTransformer`. While I was fixing that
example, I thought I'd do an editing pass on the rest of the content to
correct some grammar, and add some details I thought would be useful
based on my experience using this library.